requirements.txt
astroid==2.3.2
certifi==2019.9.11
chardet==3.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pylint==2.4.3
python-decouple==3.1
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
